rename reverse_sentence.rb & reverse_sentence_test.rb
