+++ test/test_helper.rb
@@ -4,7 +4,6 @@ require_relative '../lib/reverse_sentence'
 
 Minitest::Reporters.use! Minitest::Reporters::SpecReporter.new
 
-require_relative "../lib/reverse_sentence_option_2"
-require_relative "../lib/reverse_sentence_option_1"
+require_relative "../lib/reverse_sentence"
 require_relative "../lib/sort_by_length"
